automation/services/mina-bp-stats/leaderboard-bot/requirements.txt
Flask-SQLAlchemy==2.4.4
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
SQLAlchemy==1.4.11
Werkzeug==1.0.1
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
google-api-core==1.26.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-auth==1.27.1
google-cloud-core==1.6.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
google==3.0.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
gspread==3.7.0
httplib2==0.19.1
idna==2.10
install==1.3.4
itsdangerous==1.1.0
joblib==1.0.1
natsort==7.1.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.1.5
pd==0.0.4   
protobuf==3.15.5
psycopg2-binary==2.8.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7.2
sendgrid==6.6.0
six==1.15.0
soupsieve==2.2
starkbank-ecdsa==1.1.0
uritemplate==3.0.1
urllib3==1.26.3
networkx==2.6.3
matplotlib==3.5.1
